% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aspatial_sir}
\alias{aspatial_sir}
\title{Helper Function for Seasonal SIR Simulation}
\usage{
aspatial_sir(
  initial_pop,
  season_length,
  mortality,
  transmission,
  recovery,
  fecundity,
  abundance_threshold,
  carrying_capacity,
  season
)
}
\arguments{
\item{initial_pop}{A vector of length 6 showing the initial abundance for
each combination of stage and compartment, in the order susceptible
juvenile, susceptible adult, infected juvenile, infected adult, recovered
juvenile, recovered adult.}

\item{season_length}{The length of the season in days.}

\item{mortality}{A vector of length 6 with the mortality rates for each
stage and compartment in the season in question.}

\item{transmission}{A vector of length 6 with the transmission rates for
each stage and compartment in the season in question. Only the susceptible
entries (elements 1 and 2) are used.}

\item{recovery}{A vector of length 6 with the recovery rates for each
stage and compartment in the season in question. Only the infected entries
(elements 3 and 4) are used.}

\item{fecundity}{A vector of length 6 with the fecundity for each
reproductive segment. Only adult fecundity (element 2) is used.}

\item{abundance_threshold}{A quasi-extinction threshold below which a
population becomes extinct.}

\item{carrying_capacity}{A single numeric that indicates the carrying
capacity of the population in this season.}

\item{season}{Either "breeding" or "non-breeding."}
}
\value{
A vector of length 6 showing the abundance for each combination of
stage and compartment at the end of the season.
}
\description{
This function is an internal one that does the aspatial simulations within
one population for one timestep, for any given season. It is the two-stage
(juvenile, adult) SIR analogue of \code{\link[=aspatial_siri]{aspatial_siri()}}, in which recovery
confers permanent immunity.
}
\examples{
aspatial_sir(
 initial_pop = c(50000, 50000, 1, 0, 0, 0),
 season_length = 100,
 mortality = c(0.004, 0, 0.00505, 0.00105, 0.004, 0),
 fecundity = c(0, 15/182, 0, 15/182, 0, 15/182),
 transmission = c(0.00002, 0.00001, 0, 0, 0, 0),
 recovery = c(0, 0, 0.05714286, 0.05714286, 0, 0),
 carrying_capacity = 150000,
 abundance_threshold = 10,
 season = "breeding"
)
}
