% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspatial_siri_seasons.R
\name{aspatial_siri}
\alias{aspatial_siri}
\title{Helper function: SIRI seasonal simulator}
\usage{
aspatial_siri(
  initial_pop,
  carrying_capacity,
  abundance_threshold,
  mortality,
  transmission,
  recovery,
  fecundity,
  season
)
}
\arguments{
\item{initial_pop}{A vector of length 8 showing the initial abundance for
each combination of stage and compartment.}

\item{carrying_capacity}{A single numeric that indicates the carrying
capacity of the population in this season.}

\item{abundance_threshold}{A quasi-extinction threshold below which a
population becomes extinct.}

\item{mortality}{A vector of length 8 with the mortality rates for each
stage and compartment in the season in question.}

\item{transmission}{A vector of length 4 with the transmission rates for each
susceptible/recovered stage in the season in question.}

\item{recovery}{A vector of length 4 with the transmission rates for each
susceptible/recovered stage in the season in question.}

\item{fecundity}{Only necessary when \code{season = "breeding"} (see below).
Default NULL. A single numeric with the daily fecundity of adults.}

\item{season}{Either "breeding" or "non-breeding."}
}
\value{
A vector of length 8 showing the abundance for each combination of
stage and compartment at the end of the season.
}
\description{
Helper function: SIRI seasonal simulator
}
