% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aspatial_siri_prep}
\alias{aspatial_siri_prep}
\title{Prep Inputs for Aspatial SIRI Model Functions}
\usage{
aspatial_siri_prep(
  df,
  init,
  f = c("siri_model_summer", "siri_model_winter", "siri_model_year"),
  ...
)
}
\arguments{
\item{df}{A single row of a data frame generated by \code{simulation_input_df}.}

\item{init}{Represents the state of the population at the start of the simulation.
A list with named elements, each one a numeric of length 1:
\describe{
\item{\code{Sj}}{Susceptible juveniles.}
\item{\code{Sa}}{Susceptible adults.}
\item{\code{I1j}}{Juveniles infected for the first time.}
\item{\code{I1a}}{Adults infected for the first time.}
\item{\code{Rj}}{Recovered juveniles.}
\item{\code{Ra}}{Recovered adults.}
\item{\code{I2j}}{Juveniles infected for the second+ time.}
\item{\code{I2a}}{Adults infected for the second+ time.}
}}

\item{f}{A string indicating whether inputs are being prepped for the \code{siri_model_summer},
\code{siri_model_winter}, or \code{siri_model_year} function.}

\item{...}{A placeholder for future code improvements.}
}
\value{
A named list of all arguments needed for the selected function. Can be passed directly to that
function using \code{purrr::lift}.
}
\description{
This function converts the data frames generated by the Latin hypercube sampling of my prior
distributions into appropriate inputs for the \code{siri_model_summer}, \code{siri_model_winter}, and
\code{siri_model_year} functions.
}
