% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_transformation.R
\name{disease_transformation}
\alias{disease_transformation}
\title{Nested functions for a user-defined transformation of a population affected
by a disease outbreak.}
\usage{
disease_transformation(
  replicates,
  time_steps,
  seasons,
  populations,
  demographic_stochasticity,
  density_stages,
  abundance_threshold,
  mortality,
  mortality_unit,
  fecundity,
  fecundity_unit,
  fecundity_mask,
  transmission,
  transmission_unit,
  transmission_mask,
  recovery,
  recovery_unit,
  recovery_mask,
  transformation,
  simulator,
  name = "transformation"
)
}
\arguments{
\item{replicates}{Number of replicate simulation runs.}

\item{time_steps}{Number of simulation time steps.}

\item{seasons}{Number of seasons.}

\item{populations}{Number of populations.}

\item{demographic_stochasticity}{Boolean for optionally choosing demographic
stochasticity for the transformation.}

\item{density_stages}{Array of booleans or numeric (0,1) for each stage to
indicate which stages are affected by density.}

\item{abundance_threshold}{A quasi-extinction threshold below which a
population becomes extinct.}

\item{mortality}{A vector of mortality rates, one for each
combination of stages and compartments. Assumed by default to be daily
mortality rates unless indicated otherwise (see below).}

\item{mortality_unit}{A vector indicating whether mortality rates are
daily or seasonal. 1 indicates seasonal, 0 indicates daily. Default: all 0.
A list of vectors may be provided if this varies by season.}

\item{fecundity}{A vector of fecundity rates, one for each
combination of stages and compartments for which fecundity applies (see
\code{fecundity_mask} below).}

\item{fecundity_unit}{A vector indicating whether mortality rates are
daily or seasonal. 1 indicates seasonal, 0 indicates daily. Default: all 0.
A list of vectors may be provided if this varies by season.}

\item{fecundity_mask}{A vector indicating which stages and
compartments reproduce.}

\item{transmission}{A vector of transmission rates, one for each
combination of stages and compartment for which transmission applies (see
\code{transmission_mask} below.}

\item{transmission_unit}{A vector indicating whether mortality rates
are daily or seasonal. 1 indicates seasonal, 0 indicates daily.}

\item{transmission_mask}{A vector indicating which stages and
compartments are subject to transmission (i.e., classes susceptible to
infection.) Must be the same length as \code{compartments}.}

\item{recovery}{A vector of recovery rates, one for each
combination of stages and compartment for which recovery applies (see
\code{recovery_mask} below.) If recovery varies by season, a list of
recovery vectors the same length as \code{seasons} may be provided instead.}

\item{recovery_unit}{A vector
indicating whether mortality rates are daily or seasonal. 1 indicates
seasonal, 0 indicates daily.}

\item{recovery_mask}{A vector indicating which compartments are subject to
recovery (i.e., infected classes that can recover.) Must be the same length as
\code{compartments}.}

\item{transformation}{A user-defined function (optionally nested in a list
with additional attributes) for performing transformation:
\code{function(params)}, where \emph{params} is a list passed to the function
containing:
\describe{
\item{\code{replicates}}{Number of replicate simulation runs.}
\item{\code{time_steps}}{Number of simulation time steps.}
\item{\code{years_per_step}}{Number of years per time step.}
\item{\code{populations}}{Number of populations.}
\item{\code{stages}}{Number of life cycle stages.}
\item{\code{compartments}}{Number of disease compartments (e.g., 3 for a
SIR model).}
\item{\code{demographic_stochasticity}}{Boolean for optionally choosing
demographic stochasticity for the transformation.}
\item{\code{density_stages}}{Array of booleans or numeric (0,1) for each
stage to indicate which stages are affected by density.}
\item{\code{abundance_threshold}}{A quasi-extinction threshold below
which a population becomes extinct.}
\item{\code{mortality}}{A vector of mortality rates, one for each
combination of stages and compartments.}
\item{\code{mortality_unit}}{A vector indicating whether mortality rates
are daily or seasonal. 1 indicates seasonal, 0 indicates daily.}
\item{\code{fecundity}}{A vector of fecundity rates, one for each
combination of stages and compartments for which fecundity applies.}
\item{\code{fecundity_unit}}{A vector indicating whether fecundity rates
are daily or seasonal. 1 indicates seasonal, 0 indicates daily.}
\item{\code{fecundity_mask}}{A vector indicating which stages and
compartments reproduce.}
\item{\code{transmission}}{A vector of transmission rates, one for each
combination of stages and compartment for which transmission applies (see
\code{transmission_mask} below.}
\item{\code{transmission_unit}}{A vector indicating whether mortality
rates are daily or seasonal. 1 indicates seasonal, 0 indicates daily.}
\item{\code{transmission_mask}}{A vector indicating which stages and
compartments are subject to transmission (i.e., classes susceptible to
infection.)}
\item{\code{recovery}}{A vector of recovery rates, one for each
combination of stages and compartment for which recovery applies (see
\code{recovery_mask} below.)}
\item{\code{recovery_unit}}{A vector indicating whether mortality rates
are daily or seasonal. 1 indicates seasonal, 0 indicates daily.}
\item{\code{recovery_mask}}{A vector indicating which compartments are
subject to recovery (i.e., infected classes that can recover.)}
\item{\code{r}}{Simulation replicate.}
\item{\code{tm}}{Simulation time step.}
\item{\code{carrying_capacity}}{Array of carrying capacity values for
each population at time step.}
\item{\code{breeding_season_length}}{Array of breeding season lengths in
days for each population at time step.}
\item{\code{segment_abundance}}{Matrix of (current) abundance for each
stage-compartment combo (rows) and population (columns) at time step.}
\item{\code{occupied_indices}}{Array of indices for populations occupied
at (current) time step.}
\item{\code{simulator}}{\code{\link{SimulatorReference}} object with
dynamically accessible \emph{attached} and \emph{results} lists.}
\item{\code{additional attributes}}{Additional attributes when the
transformation is optionally nested in a list.}
}
returns a transformed stage abundance matrix (or a list with stage
abundance and carrying capacity)}

\item{simulator}{\code{\link{SimulatorReference}} object with dynamically
accessible \emph{attached} and \emph{results} lists.}

\item{name}{Optional name for the transformation (default is
"transformation").}

\item{compartments}{Number of disease compartments (e.g., 3 for a SIR model).
Default: 1.}
}
\value{
Abundance (and capacity) transformation function:
\code{function(r, tm, carrying_capacity, segment_abundance,
occupied_indices)}, where:
\describe{
\item{\code{r}}{Simulation replicate.}
\item{\code{tm}}{Simulation time step.}
\item{\code{carrying_capacity}}{Array of carrying capacity values for
each population at time step.}
\item{\code{segment_abundance}}{Matrix of abundance for each
stage-compartment combo (rows) and population (columns) at time step.}
\item{\code{occupied_indices}}{Array of indices for populations occupied
at time step.}
\item{\code{returns}}{List with transformed stage abundance matrix (and
optionally carrying capacity).}
}
}
\description{
Modular functions for the disease simulator for performing a transformation
of a population across stages and disease compartments (and optionally
carrying capacity) at a specified time step via a user-defined function.
}
