% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_transitions.R
\name{disease_transitions}
\alias{disease_transitions}
\title{Nested functions for stage- and compartment-based population transitions in
an outbreak.}
\usage{
disease_transitions(stages, compartments)
}
\arguments{
\item{stages}{Number of life cycle stages.}

\item{compartments}{Number of disease compartments.}
}
\value{
Transition calculation function that takes as input:
\describe{
\item{\code{segment_abundance}}{Matrix of (current) abundance for each
stage-compartment combo (rows) and population (columns) at time step.}
\item{\code{occupied_indices}}{Array of indices for populations occupied
at (current) time step.}
}
}
\description{
Modular functions for the disease simulator to transition populations
between stages and within disease compartments.
}
