% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finch_region}
\alias{finch_region}
\title{Study region for house finch conjunctivitis simulations}
\format{
A raster object with 1 layer and 17066 cells defining the spatial
scope of a simulation.
}
\usage{
finch_region
}
\description{
This is a raster that defines the spatial extent and resolution of the house
finch conjunctivitis example in the vignette. It encompasses terrestrial
North America from southern Mexico to southern Canada. It has a resolution of
46.375 by 46.375 kilometers, adding up to 6,355 possible populations of
house finches. The projected coordinate system for the study region is
Albers equal-area conic, with a reference longitude of 94.5 W and standard
parallels at 21.5 N and 47.5 N.
}
\keyword{datasets}
