% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitat_suitability}
\alias{habitat_suitability}
\title{House finch habitat suitability}
\format{
A RasterStack with 17066 cells and 23 layers.
}
\usage{
habitat_suitability
}
\description{
This is a RasterStack containing data on habitat suitability for
the house finch in North America from 1994 to 2016. This habitat suitability
stack was generated using a species distribution model. The predictors for
the SDM were the 12 bioclimatic variables, plus an urbanization index
(proportion of each grid cell with urban land use.) The occurrences for the
SDM came from quality-checked GBIF records.
}
\details{
The raster has the same resolution and projection as the \link{finch_region}.
Habitat suitability ranges from 0 to 1, with 1 being the most suitable.
}
\keyword{datasets}
