% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{initial_abundance}
\alias{initial_abundance}
\title{Initial house finch abundance}
\format{
A numeric matrix with 8 rows and 6355 columns.
}
\usage{
initial_abundance
}
\description{
This numeric matrix contains initial abundances of house finches in 1994 for
the house finch conjunctivitis vignette. These abundances were themselves
simulated using \code{epizootic} and do not represent empirical estimates of house
finch abundance in 1994. The matrix has 6,355 columns, one for each
population in the \link{finch_region}, and 8 rows, one for each combination of
life cycle stage and disease compartment (row 1: susceptible juveniles, row
2: susceptible adults, row 3: juveniles infected for the first time, row 4:
adults infected for the first time, row 5: recovered juveniles, row 6:
recovered adults, row 7: re-infected juveniles, row 8: re-infected adults.)
This initial abundance matrix has only zeroes in rows 3-8 because the disease
has not yet broken out at the start of the simulation.
}
\keyword{datasets}
