% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_raster}
\alias{interpolate_raster}
\title{Interpolation of Missing Timesteps in a Raster Stack}
\usage{
interpolate_raster(
  raster_stack,
  source_time,
  target_time,
  time_label,
  ...,
  method = "linear"
)
}
\arguments{
\item{raster_stack}{A \code{terra::SpatRaster} with at least two layers.}

\item{source_time}{A numeric vector indicating the timesteps of the \code{raster_stack}, whatever those
may be, e.g., \code{c(1940, 1945)}.}

\item{target_time}{A numeric vector indicating the series of timesteps desired in the output, e.g.,
\verb{c(1940, 1941, 1942. 1943, 1944, 1945)}.}

\item{time_label}{\code{terra::SpatRaster}s do not allow raw numbers as names of raster layers. Therefore
a string is needed to place before the timestep number, e.g., "BP", "BCE".}

\item{method}{Interpolation method passed on to \code{terra::approximate}. Default is "linear";
alternative is "constant", which implements a step function.}
}
\value{
A \code{terra::SpatRaster} with as many layers as the length of \code{target_time}.
}
\description{
Interpolates missing time points in a raster stack with layers for different time points. Intended
for \code{terra::SpatRaster} use only.
}
