% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspatial_siri_seasons.R
\name{siri_model_summer}
\alias{siri_model_summer}
\title{Simulate a \emph{Mycoplasma gallisepticum} Outbreak in the Breeding Season}
\usage{
siri_model_summer(initial_vector, parms, ...)
}
\arguments{
\item{initial_vector}{Represents the state of the population at the start of the breeding season.
A list with named elements, each one a numeric of length 1:
\describe{
\item{\code{Sj}}{Susceptible juveniles.}
\item{\code{Sa}}{Susceptible adults.}
\item{\code{I1j}}{Juveniles infected for the first time.}
\item{\code{I1a}}{Adults infected for the first time.}
\item{\code{Rj}}{Recovered juveniles.}
\item{\code{Ra}}{Recovered adults.}
\item{\code{I2j}}{Juveniles infected for the second+ time.}
\item{\code{I2a}}{Adults infected for the second+ time.}
}}

\item{parms}{A vector of simulation parameters:
\describe{
\item{\code{season_length}}{Length of the breeding season in days.}
\item{\code{carrying_capacity}}{Carrying capacity of finches in the population.}
\item{\code{birth}}{Daily fecundity.}
\item{\code{beta_Sj}}{Daily transmission for susceptible juveniles in breeding season.}
\item{\code{beta_Sa}}{Daily transmission for susceptible adults in breeding season.}
\item{\code{beta_Rj}}{Daily transmission for recovered juveniles in breeding season.}
\item{\code{beta_Ra}}{Daily transmission for recovered adults in breeding season.}
\item{\code{mortality_Sj}}{Daily mortality of susceptible juveniles in breeding season.}
\item{\code{mortality_Sa}}{Daily mortality of susceptible adults in breeding season.}
\item{\code{mortality_I1j}}{Daily mortality of juveniles infected for the first time.}
\item{\code{mortality_I1a}}{Daily mortality of adults infected for the first time.}
\item{\code{mortality_I2j}}{Daily mortality of juveniles infected for the second+ time.}
\item{\code{mortality_I2a}}{Daily mortality of adults infected for the second+ time.}
\item{\code{recovery_I1}}{Daily recovery rate of birds infected for the first time.}
\item{\code{recovery_I2}}{Daily recover rate of birds infected for the second+ time.}
}}

\item{...}{Does nothing. A placeholder for future code improvements.}
}
\value{
A data frame where each row is a day of the simulation and each column is one of the
population states detailed in \code{initial_vector}.
}
\description{
Simulate a \emph{Mycoplasma gallisepticum} outbreak during the breeding season day-by-day in a
population of house finches (\emph{Haemorhous mexicanus}). Uses a SIRI model (Susceptible-Infected 1-
Recovered-Infected 2+) and includes demographic stochasticity in fecundity, mortality, and
infection.
}
\details{
The function can also handle the case in which there are no infected individuals. The principal
difference between this function and the one for simulating an outbreak in the non-breeding season
is that this one includes fecundity.
}
