% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspatial_siri_seasons.R
\name{siri_model_year}
\alias{siri_model_year}
\title{Simulate a \emph{Mycoplasma gallisepticum} Outbreak Over a Year}
\usage{
siri_model_year(initial_vector, summer_params, winter_params, ...)
}
\arguments{
\item{initial_vector}{Represents the state of the population at the start of the breeding season.
A list with named elements, each one a numeric of length 1:
\describe{
\item{\code{Sj}}{Susceptible juveniles.}
\item{\code{Sa}}{Susceptible adults.}
\item{\code{I1j}}{Juveniles infected for the first time.}
\item{\code{I1a}}{Adults infected for the first time.}
\item{\code{Rj}}{Recovered juveniles.}
\item{\code{Ra}}{Recovered adults.}
\item{\code{I2j}}{Juveniles infected for the second+ time.}
\item{\code{I2a}}{Adults infected for the second+ time.}
}}

\item{summer_params}{A vector of simulation parameters defined for the breeding season:
\describe{
\item{\code{season_length}}{Length of the breeding season in days.}
\item{\code{carrying_capacity}}{Carrying capacity of finches in the population.}
\item{\code{birth}}{Daily fecundity.}
\item{\code{beta_Sj}}{Daily transmission for susceptible juveniles in breeding season.}
\item{\code{beta_Sa}}{Daily transmission for susceptible adults in breeding season.}
\item{\code{beta_Rj}}{Daily transmission for recovered juveniles in breeding season.}
\item{\code{beta_Ra}}{Daily transmission for recovered adults in breeding season.}
\item{\code{mortality_Sj}}{Daily mortality of susceptible juveniles in breeding season.}
\item{\code{mortality_Sa}}{Daily mortality of susceptible adults in breeding season.}
\item{\code{mortality_I1j}}{Daily mortality of juveniles infected for the first time.}
\item{\code{mortality_I1a}}{Daily mortality of adults infected for the first time.}
\item{\code{mortality_I2j}}{Daily mortality of juveniles infected for the second+ time.}
\item{\code{mortality_I2a}}{Daily mortality of adults infected for the second+ time.}
\item{\code{recovery_I1}}{Daily recovery rate of birds infected for the first time.}
\item{\code{recovery_I2}}{Daily recovery rate of birds infected for the second+ time.}
}}

\item{winter_params}{A vector of simulation parameters defined for the non-breeding season:
\describe{
\item{\code{season_length}}{Length of the non-breeding season in days.}
\item{\code{carrying_capacity}}{Carrying capacity of finches in the population.}
\item{\code{beta_Sj}}{Daily transmission for susceptible juveniles in non-breeding season.}
\item{\code{beta_Sa}}{Daily transmission for susceptible adults in non-breeding season.}
\item{\code{beta_Rj}}{Daily transmission for recovered juveniles in non-breeding season.}
\item{\code{beta_Ra}}{Daily transmission for recovered adults in non-breeding season.}
\item{\code{mortality_Sj}}{Daily mortality of susceptible juveniles in non-breeding season.}
\item{\code{mortality_Sa}}{Daily mortality of susceptible adults in non-breeding season.}
\item{\code{mortality_I1j}}{Daily mortality of juveniles infected for the first time.}
\item{\code{mortality_I1a}}{Daily mortality of adults infected for the first time.}
\item{\code{mortality_I2j}}{Daily mortality of juveniles infected for the second+ time.}
\item{\code{mortality_I2a}}{Daily mortality of adults infected for the second+ time.}
\item{\code{recovery_I1}}{Daily recovery rate of birds infected for the first time.}
\item{\code{recovery_I2}}{Daily recovery rate of birds infected for the second+ time.}
}}

\item{...}{Does nothing. A placeholder for future code improvements.}
}
\value{
A data frame where each row is a day of the simulation and each column is one of the
population states detailed in \code{initial_vector}.
}
\description{
Simulate a \emph{Mycoplasma gallisepticum} outbreak day-by-day in a population of house finches
(\emph{Haemorhous mexicanus}) for 365 days, starting at the beginning of the breeding season.
Uses a SIRI model (Susceptible-Infected 1-
Recovered-Infected 2+) and includes demographic stochasticity in fecundity, mortality, and
infection. The function can also handle the case in which there are no infected individuals.
}
\details{
The responsibility is on the user to ensure that the winter and summer season lengths add up in
a reasonable way.
}
