% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aspatial_siri}
\alias{aspatial_siri}
\title{Helper Function for Seasonal SIRI Simulation}
\usage{
aspatial_siri(
  pop_list,
  mortality_list,
  transmission_list,
  recovery_list,
  fecundity_list,
  abundance_threshold,
  carrying_capacity_list,
  season
)
}
\description{
Testing out a seasonal version instead of a daily timestep version.
}
