% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_simulator.R
\name{disease_simulator}
\alias{disease_simulator}
\title{Stage-based seasonal spatially explicit SIRI model of the
\emph{Mycoplasma gallisepticum} outbreak in \emph{Haemorhous mexicanus}.}
\usage{
disease_simulator(inputs)
}
\value{
Selected simulation results as a nested list summarized (mean, sd,
min, max) across multiple replicates (default), or 2-3D arrays including
results for each replicate:
\describe{
\item{\code{abundance}}{Matrix or 3D array of simulation abundance:
\emph{populations} rows by \emph{time_steps} columns (by \emph{replicates}
deep).}
\item{\code{abundance_stages}}{List of matrices or 3D arrays of simulation
abundance for unique stage-compartment combinations when present: each
\emph{populations} rows by \emph{time_steps} columns (by \emph{replicates}
deep).}
\item{\code{all$abundance}}{Array or matrix of total abundance across
populations: \emph{time_steps} (rows by \emph{replicates} columns).}
\item{\code{all$abundance_stages}}{List of arrays or matrices of total
abundance across populations for unique stage-compartment combinations when
present: each \emph{time_steps} (rows by \emph{replicates} columns).}
\item{\code{all$ema}}{Array of expected minimum abundance at each time step
(averaged across replicates).}
\item{\code{extirpation}}{Array or matrix of extirpation times:
\emph{populations} (rows by \emph{replicates} columns).}
\item{\code{all$extirpation}}{Array of extirpation time across populations
for each replicate.}
\item{\code{all$extinction_location}}{The weighted centroid of cells
occupied in the time-step prior to the extirpation of all populations
(if it occurred) for each replicate.}
\item{\code{harvested}}{Matrix or 3D array of individuals harvested:
\emph{populations} rows by \emph{time_steps} columns (by \emph{replicates}
deep).}
\item{\code{harvested_stages}}{List of matrices or 3D arrays of individuals
harvested for unique stage-compartment combinations when present: each
\emph{populations} rows by \emph{time_steps} columns (by \emph{replicates}
deep).}
\item{\code{all$harvested}}{Array or matrix of individuals harvested across
populations: \emph{time_steps} (rows by \emph{replicates} columns).}
\item{\code{all$harvested_stages}}{List of arrays or matrices of individuals
harvested across populations for unique stage-compartment combinations when
present: each \emph{time_steps} (rows by \emph{replicates} columns).}
\item{\code{all$occupancy}}{Array or matrix of the number of populations
occupied at each time-step: \emph{time_steps} (rows by \emph{replicates}
columns).}
\item{\code{additional results}}{Additional results may be attached via
user-defined functions (using \code{params$simulator$results}).}
}
}
\description{
Simulates a stage-based demographic population model and returns simulation
results across multiple replicate runs. Processes run at each simulation
time-step include:
\enumerate{
\item Density dependence calculations (ceiling, logistic, or user-defined)
\item Environmental stochasticity calculations
\item Stage transition (stochastic) calculations
\item Population growth/decline calculations
\item Disease outbreak according to a SIRI model
\item Dispersal calculations (default or user-defined)
\item Results collection
}
This simulator is not ready for use in any system other than the \emph{Mycoplasma
gallisepticum} outbreak in the house finch. Note that the breeding season is
always treated as the first season.
}
