% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epizootic-package.R
\docType{package}
\name{epizootic-package}
\alias{epizootic}
\alias{epizootic-package}
\title{epizootic: Spatially Explicit Population Models of Disease Transmission in Wildlife}
\description{
This extension of the pattern-oriented modeling framework of the 'poems' package provides a collection of modules and functions customized for modeling disease transmission on a population scale in a spatiotemporally explicit manner. This includes seasonal time steps, dispersal functions that track disease state of dispersers, results objects that store disease states, and a population simulator that includes disease dynamics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/viralemergence/epizootic}
  \item Report bugs at \url{https://github.com/viralemergence/epizootic/issues}
}

}
\author{
\strong{Maintainer}: July Pilowsky \email{pilowskyj@caryinstitute.org} (\href{https://orcid.org/0000-0002-6376-2585}{ORCID})

Other contributors:
\itemize{
  \item National Science Foundation Biology Integration Institute 2213854 [funder]
}

}
\keyword{internal}
